Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv>=0.19.0
openai==1.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt>=3.2.0
pydantic==2.4.2
httpx>=0.24.1
pydantic-settings>=2.0.3
aiohttp==3.9.1
certifi>=2024.2.2
